+++ NEWS.md
@@ -13,6 +13,7 @@ more than two objects thanks to Jouni Helske (#204)
 * drop `"draws"` and `"draws_df"` classes from `draws_df` objects if meta data
 columns are removed by a `dplyr` operation (#202)
 * fix output of `print.draws_df()` on objects with unrepaired draws (#217)
+* ensure `variance()` works properly with `summarise_draws()` (#219)
 
 
 # posterior 1.1.0